backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# Database Drivers
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
pyodbc==5.0.1
snowflake-connector-python==3.6.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# AI Integration
anthropic==0.18.1

# Utilities
python-dateutil==2.8.2
